requirements.txt
streamlit
st-supabase-connection
supabase
pandas
python-dotenv
stravaio